fix: block events count
included all records for events rather than filter records according to extrinsic index.
abstracted latest block fetch into a custom hook for snapshot and block list components
provided a function to obtain and format block time from block data.
